+++ src/enums/stash-commands.enum.ts
@@ -1,0 +1,6 @@
+export enum StashCommands {
+    show = 'scm-local-stashes.stash.show',
+    apply = 'scm-local-stashes.stash.apply',
+    pop = 'scm-local-stashes.stash.pop',
+    drop = 'scm-local-stashes.stash.drop'
+}