+++ src/enums/stash-global-commands.enum.ts
@@ -1,0 +1,6 @@
+export enum StashGlobalCommands {
+    refresh = 'scm-local-stashes.refresh',
+    create = 'scm-local-stashes.create',
+    createIncludeUntracked = 'scm-local-stashes.create-include-untracked',
+    createForRepo = 'scm-local-stashes.repo.create'
+}