+++ src/enums/tree-node-context.enum.ts
@@ -3,5 +3,6 @@ export enum TreeNodeContext {
     branchWithUpstream = 'branchWithUpstream',
     activeBranch = 'activeBranch',
     activeBranchWithUpstream = 'activeBranchWithUpstream',
+    stash = 'stash',
     repo = 'repo'
 }