+++ src/extension.ts
@@ -1,7 +1,12 @@
 import * as vscode from 'vscode';
 import { BranchSwitcher } from './branches';
+import { StashManager } from './stashes';
+import { Git } from './git';
 
 export function activate(context: vscode.ExtensionContext) {
-    new BranchSwitcher(context);
-}
+    const git = new Git();
+    context.subscriptions.push(git);
 
+    new BranchSwitcher(context, git);
+    new StashManager(context, git);
+}